+++ RasAsiVer1/External_Packeg/emojilist.py
@@ -1,0 +1,6 @@
+ej = {
+    'слон': '\U0001F418',
+    'gраф': '\U0001F992',
+    'молния': '\U000026A1',
+    'хлопушка': '\U0001F389'
+}
